# Use multiversion to speed up nonnull_sum/null_sum for cpus with AVX