--- a/src/compute/aggregate/mod.rs
+++ b/src/compute/aggregate/mod.rs
@@ -221,11 +221,9 @@ fn null_sum<T: NativeType + AddAssign + Sum>(values: &[T], bitmap: &Bitmap) -> T
         .zip(iter)
         .fold(T::new_simd(), |mut acc, (chunk, validity_chunk)| {
             let chunk = T::from_slice(chunk);
-            let mut iter = BitChunkIter::new(validity_chunk);
-            for i in 0..T::LANES {
-                if iter.next().unwrap() {
-                    acc[i] += chunk[i];
-                }
+            let iter = BitChunkIter::new(validity_chunk);
+            for (i, b) in (0..T::LANES).zip(iter) {
+                acc[i] += if b { chunk[i] } else { T::default() };
             }
             acc
         });