* Use multiversion
* Make inner loop easier to vectorize

```
Benchmarking add f32: Collecting 100 samples in estimated 5.0181 s (146k iterati                                                                                add f32                 time:   [33.429 us 33.478 us 33.533 us]
                        change: [-36.514% -36.290% -36.071%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

Benchmarking add nulls f32: Collecting 100 samples in estimated 5.0523 s (217k i                                                                                add nulls f32           time:   [22.802 us 22.814 us 22.829 us]
                        change: [-85.562% -85.523% -85.484%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe
  ```