@Dandandan 2021-04-23
@jorgecarleitao now with a >5x improvement to `null_sum`!